Move Scoring strats to their own namespaces
