+++ PlayerRank/PlayerRank.csproj
@@ -39,15 +39,15 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Scoring\Discard.cs" />
+    <Compile Include="Scoring\Elo\EloScoringStrategy.cs" />
     <Compile Include="Scoring\IScoringStrategy.cs" />
     <Compile Include="League.cs" />
     <Compile Include="PlayerScore.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Game.cs" />
-    <Compile Include="Scoring\EloScoringStrategy.cs" />
-    <Compile Include="Scoring\LowestPointsStrategy.cs" />
-    <Compile Include="Scoring\SimpleScoringStrategy.cs" />
+    <Compile Include="Scoring\LowestPoints\LowestPointsStrategy.cs" />
+    <Compile Include="Scoring\LowestPoints\DiscardPolicy.cs" />
+    <Compile Include="Scoring\Simple\SimpleScoringStrategy.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
